Update the observer to collect prometheus metrics directly
We would like to be able to determine if soak tests are erratic automatically,
REF #10606. Also in #10627 we find that while it's possible to export
cummulative bytes written from the kube it's kind of a crummy experience
overall.

This commit starts to address both by adjusting the observer to natively
understand prometheus, allowing us to export mostly-raw data from the kube for
later analysis.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>